move(embedded): into /feature
