+++ server/src/feature/embeddedLanguages/languageModes.ts
@@ -13,16 +13,16 @@ import {
   TextDocument,
 } from "vscode-html-languageservice";
 import { TextDocumentPositionParams } from "vscode-languageserver";
-import { DefinitionResult } from "../feature/definition/getDefinition";
-import { getAttributeInterpolationMode } from "../modes/getAttributeInterpolationMode";
-import { getAttributeMode } from "../modes/getAttributeMode";
-import { getAureliaHtmlMode } from "../modes/getAureliaHtmlMode";
-import { getCustomElementMode } from "../modes/getCustomElementMode";
-import { getRepeatForMode } from "../modes/getRepeatForMode";
-import { getTextInterpolationMode } from "../modes/getTextInterpolationMode";
-import { getValueConverterMode } from "../modes/getValueConverterMode";
-import { aureliaProgram } from "../viewModel/AureliaProgram";
-import { AureliaCompletionItem } from "../virtual/virtualCompletion/virtualCompletion";
+import { DefinitionResult } from "../definition/getDefinition";
+import { getAttributeInterpolationMode } from "../../modes/getAttributeInterpolationMode";
+import { getAttributeMode } from "../../modes/getAttributeMode";
+import { getAureliaHtmlMode } from "../../modes/getAureliaHtmlMode";
+import { getCustomElementMode } from "../../modes/getCustomElementMode";
+import { getRepeatForMode } from "../../modes/getRepeatForMode";
+import { getTextInterpolationMode } from "../../modes/getTextInterpolationMode";
+import { getValueConverterMode } from "../../modes/getValueConverterMode";
+import { aureliaProgram } from "../../viewModel/AureliaProgram";
+import { AureliaCompletionItem } from "../../virtual/virtualCompletion/virtualCompletion";
 import {
   HTMLDocumentRegions,
   aureliaLanguageId,
