+++ server/src/modes/getCustomElementMode.ts
@@ -1,18 +1,18 @@
-import { ViewRegionInfo } from "./../embeddedLanguages/embeddedSupport";
+import { ViewRegionInfo } from "../feature/embeddedLanguages/embeddedSupport";
 import * as path from "path";
 import {
   CustomElementRegionData,
   parseDocumentRegions,
   ViewRegionType,
-} from "../embeddedLanguages/embeddedSupport";
+} from "../feature/embeddedLanguages/embeddedSupport";
 import { TextDocumentPositionParams } from "vscode-languageserver";
-import { HTMLDocumentRegions } from "../embeddedLanguages/embeddedSupport";
-import { LanguageModelCache } from "../embeddedLanguages/languageModelCache";
+import { HTMLDocumentRegions } from "../feature/embeddedLanguages/embeddedSupport";
+import { LanguageModelCache } from "../feature/embeddedLanguages/languageModelCache";
 import {
   LanguageMode,
   Position,
   TextDocument,
-} from "../embeddedLanguages/languageModes";
+} from "../feature/embeddedLanguages/languageModes";
 import { getAureliaVirtualCompletions } from "../virtual/virtualCompletion/virtualCompletion";
 import { getBindablesCompletion } from "../feature/completions/completions";
 import { aureliaProgram } from "../viewModel/AureliaProgram";
